pyproject.toml
[project]
name = "interpolation"
version = "0.1.0"
description = "IITP RAS Python class"
authors = [
    {name = "Andrey Kirkicha",email = "kirkicha.aa@phystech.edu"}
]
homepage = "https://github.com/andreykirkicha/Python_IITP_RAS.git"
repository = "https://github.com/andreykirkicha/Python_IITP_RAS.git"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "ruff (>=0.11.0,<0.12.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
example = "example.console:usage"

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "SIM", "I"]
ignore = ["F401"]
fixable = ["ALL"]
unfixable = []

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
myst-parser = "^4.0.1"
sphinx-rtd-theme = "^3.0.2"
sphinx-click = "^6.0.0"
sphinx-autoapi = "^3.6.0"
linkify-it-py = "^2.0.3"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
